Separate out bingo command
