+++ .gitignore
@@ -3,3 +3,5 @@
 db.sqlite
 db.sqlite-shm
 db.sqlite-wal
+
+.idea
