+++ src/commands/bingo.rs
@@ -1,0 +1,57 @@
+use poise::CreateReply;
+use serenity::all::CreateAttachment;
+use crate::commands::{CmdContext, Error};
+
+async fn send_bingo_image(ctx: &CmdContext<'_>, image: &[u8]) -> Result<(), Error> {
+    let attachment = CreateAttachment::bytes(image, "bingo_squares.png");
+    let reply = CreateReply::default().attachment(attachment);
+
+    ctx.send(reply).await?;
+    Ok(())
+}
+
+// This can never be called, just needed for bingo subcommands
+#[poise::command(slash_command, subcommands("add", "status", "remove"))]
+pub async fn bingo(_ctx: CmdContext<'_>) -> Result<(), Error> {
+    Ok(())
+}
+
+/// Displays the current status of the bad ctf bingo squares
+#[poise::command(slash_command)]
+pub async fn status(ctx: CmdContext<'_>) -> Result<(), Error> {
+    let competition = crate::commands::competitions::get_competition_from_channel(&ctx).await?;
+
+    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
+
+    Ok(())
+}
+
+/// Checks off a bad ctf bingo square
+#[poise::command(slash_command)]
+pub async fn add(
+    ctx: CmdContext<'_>,
+    #[description = "The bingo square to check off"] square: BingoSquare,
+) -> Result<(), Error> {
+    let mut competition = crate::commands::competitions::get_competition_from_channel(&ctx).await?;
+    competition.bingo |= square;
+
+    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
+    ctx.data().db.update_competition(competition).await?;
+
+    Ok(())
+}
+
+/// Unmarks a bad ctf bingo square
+#[poise::command(slash_command)]
+pub async fn remove(
+    ctx: CmdContext<'_>,
+    #[description = "The bingo square to uncheck"] square: BingoSquare,
+) -> Result<(), Error> {
+    let mut competition = crate::commands::competitions::get_competition_from_channel(&ctx).await?;
+    competition.bingo.remove(square);
+
+    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
+    ctx.data().db.update_competition(competition).await?;
+
+    Ok(())
+}
