+++ src/commands/competitions.rs
@@ -31,7 +31,7 @@ pub async fn competition(
     let channel = CreateChannel::new(&name)
         .category(CTF_CATEGORY_ID)
         .position(0)
-        .topic(&format!("Channel for {name}"))
+        .topic(&format!("Channel for {name}; please check pinned message for shared credentials."))
         .execute(ctx, B01LERS_GUILD_ID)
         .await?;
 
@@ -51,9 +51,9 @@ pub async fn competition(
         name: name.clone(),
         bingo: BingoSquare::Free.into(),
     };
-    ctx.data().db.create_competiton(competition).await?;
+    ctx.data().db.create_competition(competition).await?;
 
-    ctx.say(format!("Created channel for {name}: {channel}"))
+    ctx.say(format!("Created channel for **{name}**: {channel}"))
         .await?;
 
     Ok(())
@@ -71,64 +71,4 @@ async fn get_competition_from_channel(ctx: &CmdContext<'_>) -> Result<Competitio
         .with_context(|| "Not in a competition channel")?;
 
     Ok(competition)
-}
-
-async fn send_bingo_image(ctx: &CmdContext<'_>, image: &[u8]) -> Result<(), Error> {
-    let attachment = CreateAttachment::bytes(image, "bingo_squares.png");
-
-    let reply = CreateReply::default().attachment(attachment);
-
-    ctx.send(reply).await?;
-
-    Ok(())
-}
-
-// This can never be called, just needed for bingo subcommands
-#[poise::command(slash_command, subcommands("add", "status", "remove"))]
-pub async fn bingo(_ctx: CmdContext<'_>) -> Result<(), Error> {
-    Ok(())
-}
-
-/// Displays the current status of the badctf bingo squares
-#[poise::command(slash_command)]
-pub async fn status(ctx: CmdContext<'_>) -> Result<(), Error> {
-    let competition = get_competition_from_channel(&ctx).await?;
-
-    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
-
-    Ok(())
-}
-
-/// Checks off a badctf bingo square
-#[poise::command(slash_command)]
-pub async fn add(
-    ctx: CmdContext<'_>,
-    #[description = "The bingo square to check off"] square: BingoSquare,
-) -> Result<(), Error> {
-    let mut competition = get_competition_from_channel(&ctx).await?;
-
-    competition.bingo |= square;
-
-    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
-
-    ctx.data().db.update_competition(competition).await?;
-
-    Ok(())
-}
-
-/// Unmarks a badctf bingo square
-#[poise::command(slash_command)]
-pub async fn remove(
-    ctx: CmdContext<'_>,
-    #[description = "The bingo square to uncheck"] square: BingoSquare,
-) -> Result<(), Error> {
-    let mut competition = get_competition_from_channel(&ctx).await?;
-
-    competition.bingo.remove(square);
-
-    send_bingo_image(&ctx, &competition.get_bingo_picture_png_bytes()?).await?;
-
-    ctx.data().db.update_competition(competition).await?;
-
-    Ok(())
 }
