+++ src/main.rs
@@ -2,7 +2,7 @@ mod commands;
 mod db;
 
 use std::env;
-
+use dotenvy::dotenv;
 use serenity::all::{ChannelId, ClientBuilder, GatewayIntents, GuildId};
 
 use commands::CommandContext;
@@ -14,15 +14,15 @@ const OFFICER_ROLE: &str = "officer";
 
 #[tokio::main]
 async fn main() {
-    dotenvy::dotenv().ok();
+    dotenv().ok();
 
-    let datbase_url =
+    let database_url =
         env::var("DATABASE_URL").expect("No `DATABASE_URL` environment variable specified");
 
     let discord_token =
         env::var("DISCORD_TOKEN").expect("No `DISCORD_TOKEN` environment variable specified");
 
-    let db = DbContext::connect(&datbase_url)
+    let db = DbContext::connect(&database_url)
         .await
         .expect("failed to connect to database");
 
@@ -30,14 +30,14 @@ async fn main() {
         .options(poise::FrameworkOptions {
             commands: vec![
                 commands::competitions::competition(),
-                commands::competitions::bingo(),
+                commands::bingo::bingo(),
             ],
             ..Default::default()
         })
         .setup(|ctx, _ready, framework| {
             Box::pin(async move {
                 // register the bots commands with discord api on startup
-                //poise::builtins::register_globally(ctx, &framework.options().commands).await?;
+                // poise::builtins::register_globally(ctx, &framework.options().commands).await?;
                 poise::builtins::register_in_guild(
                     ctx,
                     &framework.options().commands,
